+++ src/model/TransitionPath.ts
@@ -1,4 +1,5 @@
 import { NamedElement } from './NamedElement';
+import { TransitionKind } from './TransitionKind';
 
 /** Interface describing elements to leave and enter when traversing the transition; derived from the source and target using the TransitionType strategy. */
 export class TransitionPath {
@@ -7,5 +8,5 @@ export class TransitionPath {
 	 * @param leave The optional named element to leave when traversing a transition.
 	 * @param enter The optional set of elements to enter when traversing a transition.
 	 */
-	constructor(public readonly leave: NamedElement | undefined = undefined, public readonly enter: Array<NamedElement> | undefined = undefined) { }
+	constructor(public readonly kind: TransitionKind, public readonly leave: NamedElement | undefined = undefined, public readonly enter: Array<NamedElement> | undefined = undefined) { }
 }